LayoutTests/imported/w3c/web-platform-tests/css/css-position/position-absolute-dynamic-containing-block-expected.txt

TTT

PASS abs containing block moves from outer to intermediate
PASS abs containing block moves from intermediate to outer
PASS target is no longer absolute
PASS target becomes absolute
FAIL fixed containing block moves from outer to intermediate assert_equals: expected 310 but got 600
FAIL fixed containing block moves from intermediate to outer assert_equals: expected 210 but got 600
FAIL target is no longer fixed assert_equals: expected 310 but got 600
FAIL target becomes fixed assert_equals: expected 310 but got 600
